requirements.txt
appdirs
dotenv
pyyaml
redislite
requests
